requirements.txt
Django
djangorestframework
Pillow
opencv-python
imutils
scikit-image
torch 
torchvision 
torchaudio
matplotlib